+++ frontend/src/components/mypage/MyWritings.tsx
@@ -1,41 +1,17 @@
-import React from 'react';
-import ROUTE from '@constants/route';
-import { useNavigate } from 'react-router-dom';
-import COMMON from '@constants/common';
-import useServerSidePagination from '@query/get/useServerSidePagination';
-import { ArticleDetailType } from '@components/community/CommunityType';
+import React, { Suspense } from 'react';
+import Skeleton from '@components/community/loading/Skeleton';
 
-import * as C from '@components/mypage/style/MyPageComponent.style';
+import WritingsInner from './mypage/WritingsInner';
 import * as M from './style/MyWritings.style';
-import DeleteButton from './button/ArticleDeleteButton';
-import ModifyButton from './button/ModifyButton';
 
 function MyWritings() {
-  const { curPageItem, renderPaginationBtnOrInfinityScroll } = useServerSidePagination<ArticleDetailType>({
-    uri: '/api/profile/articles',
-    size: COMMON.PAGINATION.SIZE,
-  });
-
-  const navigate = useNavigate();
-
-  const goDetail = (article: ArticleDetailType) => {
-    navigate(`${ROUTE.COMMUNITY.BASE}/${article.articleId}`);
-  };
-
   return (
     <M.Container>
       <M.GroupContainer>
         <M.ContentsContainer>
-          {Object.values(curPageItem).map((item: ArticleDetailType, index) => (
-            <M.ContentsArea key={index}>
-              <M.ContentsTitle onClick={() => goDetail(item)}>{item.subject}</M.ContentsTitle>
-              <M.Contents onClick={() => goDetail(item)}>{item.content}</M.Contents>
-              <ModifyButton articleId={item.articleId}></ModifyButton>
-              <DeleteButton id={item.articleId}></DeleteButton>
-              <M.Divider />
-            </M.ContentsArea>
-          ))}
-          <C.PageSelectArea>{renderPaginationBtnOrInfinityScroll()}</C.PageSelectArea>
+          <Suspense fallback={<Skeleton />}>
+            <WritingsInner />
+          </Suspense>
         </M.ContentsContainer>
       </M.GroupContainer>
     </M.Container>
